fix: split runner dependency materialization
